Split out ingest code
